requirements.txt
aiogram==3.4.1
web3==6.15.1
python-dotenv==1.0.0
aiosqlite==0.19.0
apscheduler==3.10.4
cryptography==41.0.7
requests==2.31.0